Abstract more synteny functions
